Check if a photo and an entity are connected before removing or adding

+++ src/schemas/photo.rs
@@ -138,11 +138,29 @@ impl Photo {
         Ok("File deleted successfully!".to_string())
     }
 
+    async fn check_if_photo_has_entity(photo_id: i32, entity_id: i32, pool: &Pool) -> Result<bool, PoolError> {
+        let client = pool.get().await?;
+        let stmt = client.prepare("select count(*) \
+                                                    from photo_entity \
+                                                    where photo_id = $1 and entity_id = $2").await?;
+        let result = client.query_one(&stmt, &[&photo_id, &entity_id]).await?;
+        let count: i64 = result.get(0);
+
+        Ok(count > 0)
+    }
+
     pub async fn add_entity_to_photo(
         photo_id: i32,
         entity_id: i32,
         pool: &Pool,
     ) -> Result<String, PoolError> {
+        // check if photo is already associated with photo to prevent duplicates
+        let exists = Photo::check_if_photo_has_entity(photo_id, entity_id, pool).await?;
+
+        if exists {
+            return Ok("Entity is already associated with photo".to_string())
+        }
+
         let client = pool.get().await?;
         let stmt = client
             .prepare("insert into photo_entity (photo_id, entity_id) values ($1, $2)")
@@ -162,6 +180,12 @@ impl Photo {
         entity_id: i32,
         pool: &Pool,
     ) -> Result<String, PoolError> {
+        // check if photo has entity to prevent redundant removal
+        let exists = Photo::check_if_photo_has_entity(photo_id, entity_id, pool).await?;
+        if exists {
+            return Ok("Entity is not associated with photo".to_string())
+        }
+
         let client = pool.get().await?;
         let stmt = client
             .prepare("delete from photo_entity where photo_id = $1 and entity_id = $2")
